Add EC related parameters (#26)
Add the WP parameter (num_ec_pieces_per_segment) to ensure that all
network settings yield a consistent G size of 4104:

| Network      | V    | C   | WP   |
|--------------|:------:|:-----:|:------:|
| **tiny**         | 6    | 2   | 1026 |
| **small**        | 12   | 4   | 513  |
| **medium**       | 18   | 6   | 342  |
| **large**        | 36   | 12  | 171  |
| **xlarge**       | 108  | 36  | 57   |
| **2xlarge**      | 342  | 114 | 18   |
| **3xlarge**      | 684  | 228 | 9    |
| **full/toaster** | 1023 | 341 | 6    |


**Notes:** 
• In most network settings, 3 * C = V (i.e. three times the number of
cores equals the number of validators). Additionally, 2 * C * WP = G
(4104), or C * WP = 2052. For example, - In **tiny**: 2 * 1026 = 2052;
In **small**: 4 * 513 = 2052....

**Exception**: The **Full/Toaster** network setting’s V and C values are
rounded down from (V = 342, C = 1026) to (C = 341, V = 1023) in order to
satisfy the ringSet constraint of num_validators <= 1023. This results
in 341 * 6 = 2046, which is slightly less than 2052. In other words,
**Full/Toaster**'s EC are encoded in a way that shard_index of [1023,
1024 1025] are **dropped/ignored/not distributed/** to any DA.